Scope social metadata to homepage
